+++ gitmesh-repository-server/src/main/scala/com/github/takezoe/gitmesh/repository/api/CloneRequest.scala
@@ -1,0 +1,3 @@
+package com.github.takezoe.gitmesh.repository.api
+
+case class CloneRequest(nodeUrl: String, empty: Boolean)
