+++ gitmesh-repository-server/src/main/scala/com/github/takezoe/gitmesh/repository/api/Status.scala
@@ -1,0 +1,3 @@
+package com.github.takezoe.gitmesh.repository.api
+
+case class Status(url: String, diskUsage: Double, repos: Seq[String])
