+++ gitmesh-repository-server/src/main/scala/com/github/takezoe/gitmesh/repository/api/SynchronizedRequest.scala
@@ -1,0 +1,3 @@
+package com.github.takezoe.gitmesh.repository.api
+
+case class SynchronizedRequest(nodeUrl: String)
