+++ gitmesh-repository-server/src/main/scala/com/github/takezoe/gitmesh/repository/api/models.scala
@@ -1,0 +1,5 @@
+package com.github.takezoe.gitmesh.repository.api
+
+object models {
+
+}
