+++ gitmesh-repository-server/src/main/scala/com/github/takezoe/gitmesh/repository/util/syntax.scala
@@ -1,4 +1,4 @@
-package com.github.takezoe.gitmesh.repository
+package com.github.takezoe.gitmesh.repository.util
 
 object syntax {
